Conflicts:
	SConstruct

+++ include/mapnik/value_hash.hpp
@@ -2,7 +2,7 @@
  *
  * This file is part of Mapnik (c++ mapping toolkit)
  *
- * Copyright (C) 2014 Artem Pavlenko
+ * Copyright (C) 2015 Artem Pavlenko
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -51,7 +51,7 @@ struct value_hasher
 
     std::size_t operator() (value_unicode_string const& val) const
     {
-        return val.hashCode();
+        return static_cast<std::size_t>(val.hashCode());
     }
 
     template <class T>
